Fix three None-units/tensor display and comparison bugs found via real testing and docs builds; suppress an unrelated stderr warning in the docs
Quantity.__format__ (None-units)
- Pint's own machinery (notably pint.util.PrettyIPython._repr_latex_,
  invoked automatically by Jupyter/IPython whenever a quantity is
  displayed) formats with spec 'L' and similar Pint-specific format-type
  letters ('Lx', 'H', 'P', 'C', 'D', plus the '~'/'#' modifiers), which
  control how *units* are rendered. For a units=None quantity these were
  passed straight through to the bare NumPy/PyTorch magnitude, which
  doesn't understand them and raised TypeError ("unsupported format
  string"). This crashed the units.md and math.md doc pages (and any
  interactive display of a None-units quantity) any time such a
  quantity was shown as a cell's/REPL's result. Fixed by stripping
  Pint's format-type flags before formatting the magnitude, with a
  plain str() fallback for anything unanticipated.

Quantity.__eq__/__ne__ (tensor-backed, real units)
- Pint's own __eq__/__ne__ contain a "compare to zero" shortcut
  (`eq(magnitude, 0, True)`, where the trailing True means "reduce the
  elementwise result to a single bool via .all()") that assumes the
  comparison result is a type Pint's `is_duck_array_type` recognizes.
  It does not recognize a PyTorch tensor, so the reduction is skipped,
  and the un-reduced, multi-element boolean tensor then hits a bare
  Python `and`, whose multi-element Tensor.__bool__ raises "Boolean
  value of Tensor with more than one value is ambiguous". This crashed
  `==`/`!=` (and thus immlib.math.equal) between any two ordinary,
  real-units, multi-element tensor-backed quantities, regardless of
  whether either magnitude actually contained a zero--found via Noah's
  real `python -m unittest immlib.test` run (test_comparisons_tensor).
  Fixed by adding a tensor-safe bypass (_binop_tensor_bool, mirroring
  the existing None-units _binop_none_bool pattern) that compares
  magnitudes directly after the same same-units/convertible-units
  logic Pint itself uses, raising pint.DimensionalityError for
  incompatible units (consistent with how immlib.math's own comparison
  helpers already treat incompatible units). compare() (`<`/`<=`/`>`/
  `>=`) has no equivalent shortcut and was not affected.

Quantity.__repr__ (None-units)
- Pint's own __repr__ unconditionally quotes the units part of the
  repr string, which renders a units=None quantity's units as the
  literal string "'None'" (e.g. `<Quantity([...], 'None')>`),
  indistinguishable at a glance from a real unit named "None". Added a
  __repr__ override that calls Pint's own implementation and swaps the
  trailing `'None')>` for an unquoted `None)>`, leaving Pint's own
  (version-dependent) magnitude-formatting rules, and real-units
  reprs, untouched.

Tests
- Added regression tests for all three fixes above (test_quantity.py):
  __format__/_repr_latex_ display of None-units quantities, direct
  tensor-backed real-units `==`/`!=` (same units, convertible units,
  and incompatible units), and repr()'s bare-None-literal output.
- test_math.py's test_shape_combination_tensor compared a stack/
  concatenate result against a hardcoded float32 literal, but mixing a
  float64 NumPy array with a float32 tensor legitimately promotes to
  float64 (PyTorch's own type-promotion rule, per the "PyTorch
  controls dtype promotion" design invariant)--not a library bug.
  Fixed the test to compare against the actual result dtype instead of
  assuming either input's original dtype.

Docs build
- docs/immlib/_config.yml: every docs page triggers a stderr
  FutureWarning on `import immlib`, from a google-api-core dependency
  pulled in by immlib.iolib's optional Google Cloud Storage support--
  unrelated to anything any given page is actually demonstrating. Set
  `stderr_output: remove-warn` so this (and anything else routine on
  stderr) no longer clutters the rendered pages, while still surfacing
  as a Sphinx build-time warning if something on stderr ever indicates
  a real problem.

+++ src/immlib/test/math/test_math.py
@@ -289,12 +289,19 @@ class TestMath(TestCase):
             il.quant(np.array([1.0, 2.0]), 'm'),
             il.quant(torch.tensor([100.0, 200.0]), 'cm')])
         self.assertTrue(torch.is_tensor(r.m))
-        self.assertTrue(torch.allclose(r.m, torch.tensor([[1.0, 2.0], [1.0, 2.0]])))
+        # The NumPy array (float64) and the tensor (float32) legitimately
+        # promote to float64 (PyTorch's own type-promotion rule for
+        # stack/cat, per the "PyTorch controls dtype promotion" design
+        # invariant), so the expected tensor must match r.m's dtype rather
+        # than assume either input's original dtype.
+        self.assertTrue(torch.allclose(
+            r.m, torch.tensor([[1.0, 2.0], [1.0, 2.0]], dtype=r.m.dtype)))
         r = im.concatenate([
             il.quant(np.array([1.0, 2.0]), 'm'),
             il.quant(torch.tensor([100.0, 200.0]), 'cm')])
         self.assertTrue(torch.is_tensor(r.m))
-        self.assertTrue(torch.allclose(r.m, torch.tensor([1.0, 2.0, 1.0, 2.0])))
+        self.assertTrue(torch.allclose(
+            r.m, torch.tensor([1.0, 2.0, 1.0, 2.0], dtype=r.m.dtype)))
 
     # Linear algebra #############################################################
     def test_matmul(self):
